API Reference

Simple Mesh

MaterialPointGenerator.meshbuilder — Function

meshbuilder(x::T, y::T; ϵ="FP64") where T <: AbstractRange

Description:

Generate structured mesh in 2D space.

meshbuilder(x::T, y::T, z::T; precision::String="FP64") where T <: AbstractRange

Description:

Generate structured mesh in 3D space.

Polygon

MaterialPointGenerator.particle_in_polygon — Function

particle_in_polygon(px, py, polygon)

Description:

Determine whether a point is inside a polygon. Note the vertices of the polygon should be ordered in a counterclockwise manner; otherwise, it may lead to incorrect results.

(This function is generated by ChatGPT 4o - 2024.10.10)

MaterialPointGenerator.polygon2particle — Function

polygon2particle(domain, lpx, lpy)

Description:

Generate structured particles from a given polygon. Note the vertices of the polygon should be ordered in a counterclockwise manner; otherwise, it may lead to incorrect results.

lpx and lpy are the space of particles in x and y directions, respectively.

domain is the polygon, for example, domain = [0 0; 2 0; 2 1; 0 1] means generate structured particles in a rectangle area.

Polyhedron

MaterialPointGenerator.polyhedron2particle — Function

polyhedron2particle(stl_file::String, output_file, lp; verbose::Bool=false)

Description:

Convert a polyhedron (.stl) to a set of particles. The function will write the populated particles of each voxel into a .xyz file. The voxel size is defined by lp, it is suggest to be equal to the MPM background grid size. The verbose is a flag to show the time consumption of each step.

Example:

polyhedron2particle("/path/to/yout/model.stl", "/path/to/model.xyz", 0.1, verbose=true)

MaterialPointGenerator.readmsh — Function

readmsh(msh_path::String; precision="FP32")

Description:

Read the mesh file (.msh) generated by Gmsh. The function will return the vertices and faces for triangle. It's necessary to check the mesh qulity before using this function. You can do it in MeshLab or Gmsh GUI. precision can be "FP64" or "FP32" in String.

MaterialPointGenerator.trimesh_voxelize — Function

trimesh_voxelize(stl_file::String, lp)

Description:

Voxelize the given STL model with the trimesh package. The voxel size is defined by lp.

DEM

MaterialPointGenerator.dem2particle — Function

dem2particle(dem, lpz, bottom)

Description:

Generate particles from a given DEM file. dem is a coordinates Array with three columns (x, y, z). lpz is the space of particles in z direction. bottom is a Float64 value, which means the plane z = bottom.

dem2particle(dem, lpz, bottom_surf)

Description:

Generate particles from a given DEM file and a bottom surface file. dem is a coordinates Array with three columns (x, y, z). bottom_surf is a coordinates Array with three columns, but it should have the same x and y coordinates as the DEM, and the z value should be lower than the DEM. lpz is the space of particles in z direction.

MaterialPointGenerator.rasterizeDEM — Function

rasterizeDEM(lpx, lpy, dem; k=10, p=2, trimbounds=[0.0 0.0], dembounds=[0.0, 0.0])

Description:

Rasterize the DEM file to generate particles. lpx and lpy are the space of particles in x and y directions. dem is a coordinates Array with three columns (x, y, z). k is the number of nearest neighbors (10 by default), p is the power parameter (2 by default), trimbounds is the boundary of the particles, dembounds is the boundary of the DEM.

MaterialPointGenerator.IDW! — Function

IDW!(k, p, dem, idxs, ptslist, tree)

Description:

Inverse Distance Weighting (IDW) interpolation method. k is the number of nearest neighbors, p is the power parameter, dem is a coordinates Array with three columns (x, y, z), idxs is the index of the nearest neighbors, ptslist is the coordinates Array of the particles, tree is the KDTree of the DEM.

Utils

MaterialPointGenerator.savexyz — Function

savexyz(file_dir::P, pts::T) where {P <: String, T <: Array{Float64, 2}}

Description:

Save the points pts to the xyz file file_dir.

MaterialPointGenerator.readxyz — Function

readxyz(file_dir::P) where P <: String

Description:

Read the xyz file from file_dir.

MaterialPointGenerator.sortbycol — Function

sortbycol(pts, col::T) where T <: Int

Description:

Sort the points in pts according to the column col.