API Reference

Simple Mesh

MaterialPointGenerator.meshbuilder — Function

meshbuilder(x::T, y::T) where T <: AbstractRange

Description:

Generate structured mesh in 2D space.

meshbuilder(x::T, y::T, z::T) where T <: AbstractRange

Description:

Generate structured mesh in 3D space.

Polygon

MaterialPointGenerator.particle_in_polygon — Function

particle_in_polygon(px, py, polygon)

Description:

Determine whether a point is inside a polygon. Note the vertices of the polygon should be ordered in a counterclockwise manner; otherwise, it may lead to incorrect results.

(This function is generated by ChatGPT 4o - 2024.10.10)

MaterialPointGenerator.polygon2particle — Function

polygon2particle(domain, lpx, lpy)

Description:

Generate structured particles from a given polygon. Note the vertices of the polygon should be ordered in a counterclockwise manner; otherwise, it may lead to incorrect results.

lpx and lpy are the space of particles in x and y directions, respectively.

domain is the polygon, for example, domain = [0 0; 2 0; 2 1; 0 1] means generate structured particles in a rectangle area.

Polyhedron

MaterialPointGenerator.polyhedron2particle — Function

polyhedron2particle(msh_path::String, lpx, lpy, lpz, ::Val{:CPU})

Description:

Generate structured particles from a given polyhedron. Please provide .msh file that generated by Gmsh GUI/Script.

MaterialPointGenerator.particle_in_polyhedron! — Function

particle_in_polyhedron!(pts, node, tet, results)

Description:

Determine whether a point is inside a polyhedron. This function is able to run on any hardware backends, so the input need to preallocate the results array. Users should not call this function directly, please check polyhedron2particle.

MaterialPointGenerator.gmsh_mesh3D — Function

gmsh_mesh3D(msh_path::String)

Description:

Generate 3D mesh using Gmsh.

DEM

MaterialPointGenerator.dem2particle — Function

dem2particle(dem, lpz, bottom)

Description:

Generate particles from a given DEM file. dem is a coordinates Array with three columns (x, y, z). lpz is the space of particles in z direction. bottom is a Float64 value, which means the plane z = bottom.

dem2particle(dem, lpz, bottom_surf)

Description:

Generate particles from a given DEM file and a bottom surface file. dem is a coordinates Array with three columns (x, y, z). bottom_surf is a coordinates Array with three columns, but it should have the same x and y coordinates as the DEM, and the z value should be lower than the DEM. lpz is the space of particles in z direction.

MaterialPointGenerator.surrogateDEM — Function

surrogateDEM(lpx, lpy, dem; trimbounds, dembounds)

Description:

Generate particles from a given DEM file. lpx and lpy are the space of particles in x and y directions, respectively. dem is a coordinates Array with three columns (x, y, z). trimbounds is the bounding box of the particles, if not given, it will be the minimum and maximum of the DEM file. dembounds is the bounding box of the DEM file, this is for a DEM which is not a rectangle area, so dembounds represents the polygon of the DEM, the returned particles will be inside this polygon.

MaterialPointGenerator.rbf_compute! — Function

rbf_compute!(ptslist, x2, rbf_model)

Description:

Compute the value of the RBF model at the given points. ptslist is a 2D array with three columns (x, y, z). x2 is a 1D array with two columns (x, y). rbf_model is a RadialBasis model.