API Reference

Struct

MaterialPointGenerator.PointCloud — Type

PointCloud{T1<:TInt, T2<:TFloat}

Description:

Point cloud data structure. Coordinates x, y, z are parameters that must be input.

Functions

MaterialPointGenerator.read_pointcloud — Function

read_pointcloud(data_path::String; precision::Symbol=:FP64)

Description:

Read point cloud data from file. Users can also specify the precision of the data, available options are :FP64 and :FP32.

MaterialPointGenerator.write_pointcloud — Function

write_pointcloud(data::AbstractArray, output_path::String)

Description:

Write point cloud data to file. The output file can be read as .xyz file in other software.

MaterialPointGenerator.mp_generate — Function

mp_generate(data::PointCloud{T1, T2}, min_z::T2) where {T1, T2}

Description:

Generate a point set with a minimum z value. The minimum z value is specified by the user.